Add Authentication in Admin App
Admin needs a feature to add roles and users
in the system. The admin while creating the users
should be able to assign roles to them. The commit
also adds a feature which covers a user should be
able to sign in to the system and sign out.

This change
- Adds CRUD for roles in admin app.
- Adds CRUD for user and associate them with roles.
- Adds authenication for a user in the system.

[delivers #158523727]
[delivers #158280455]